Refactoring based on PR feedback
